requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
colorama==0.4.4
Django==4.0.3
django-ckeditor-5==0.0.16
django-js-asset==2.0.0
django-utils-six==2.0
Django-Verify-Email==2.0.1
isort==5.10.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
